+++ src/lib/plugins/gsvg.animateto.script.js
@@ -1,4 +1,0 @@
-import plugin from './gsvg.animateto.js';
-if (gSVG) {
-  gSVG.install(plugin);
-}