+++ src/lib/plugins/gsvg.debug.script.js
@@ -1,4 +1,0 @@
-import plugin from './gsvg.debug.js';
-if (gSVG) {
-  gSVG.install(plugin);
-}