+++ src/lib/plugins/gsvg.keep.aspect.script.js
@@ -1,4 +1,0 @@
-import plugin from './gsvg.keep.aspect.js';
-if (gSVG) {
-  gSVG.install(plugin);
-}