+++ src/lib/plugins/gsvg.observe.style.script.js
@@ -1,4 +1,0 @@
-import plugin from './gsvg.observe.style.js';
-if (gSVG) {
-  gSVG.install(plugin);
-}