+++ src/plugins/keep.aspect.script.js
@@ -1,4 +1,0 @@
-import {svgPlugin} from './keep.aspect.js';
-if (gSVG) {
-  gSVG.install(svgPlugin);
-}