+++ src/plugins/template.engine.js
@@ -13,7 +13,7 @@ const DIRECTIVES = Symbol();
 const EVENTS     = Symbol();
 const REPLACE    = Symbol();
 const UNKNOWN    = 'unknown';
-const directives = [];
+const directives = {};
 
 /**
  * Throws an error with a specified message, scope, and code context.
@@ -54,7 +54,7 @@ const throwError = (message, scope, code) => {
  * @returns {string} The normalized attribute name if found, else the original attribute name.
  */
 const normalizeAttribute = ((attributes) =>
-  (attribute) => attributes[attribute] || attribute
+    (attribute) => attributes[attribute] || attribute
 )(
   'attributeName attributeType baseFrequency calcMode clipPathUnits diffuseConstant edgeMode gradientTransform gradientUnits kernelMatrix kernelUnitLength lengthAdjust limitingConeAngle markerHeight markerUnits markerWidth maskContentUnits maskUnits numOctaves pathLength patternContentUnits patternTransform patternUnits pointsAtX pointsAtY pointsAtZ preserveAlpha preserveAspectRatio primitiveUnits refX refY requiredExtensions requiredFeatures specularConstant specularExponent spreadMethod startOffset stdDeviation stitchTiles surfaceScale systemLanguage tableValues targetX targetY textLength viewBox xChannelSelector yChannelSelector zoomAndPan'
     .split(' ')
@@ -108,7 +108,33 @@ function restoreFromComment (comment) {
 defineDirective({
   name : 'g-content',
   execute (gObject, {expression, data, evalExpression}) {
-    gObject.content(evalExpression(expression, data));
+    // gObject.content(evalExpression(expression, data));
+    debugger;
+    const context = {
+      ...data,
+      $$ : {
+        fromURL        : async (src) => {
+          const res = await fetch(src);
+          if (res.status === 200) {
+            return res.text();
+          }
+          console.warn(`Failed to load URL: ${ src } (${ res.status })`);
+        },
+        currentContent : gObject.content
+      }
+    };
+    const result  = evalExpression(expression, context);
+    const event   = new CustomEvent('load', {bubbles : true, detail : gObject});
+    const norm    = c => isUndefined(c) ? '' : c;
+    if (typeof result === 'object' && result.then) {
+      result.then(result => {
+        gObject.content(norm(result));
+        gObject.dispatchEvent(event);
+      });
+    } else {
+      gObject.content(norm(result));
+      gObject.dispatchEvent(event);
+    }
   }
 });
 
@@ -139,14 +165,15 @@ defineDirective({
   alias    : ':',
   argument : true,
   execute (gObject, {expression, argument, data, evalExpression}) {
-    argument           = normalizeAttribute(argument);
-    const context      = {
+    argument                = normalizeAttribute(argument);
+    const context           = {
       ...data,
       $$ : ['d', 'transform'].includes(argument) ?
         gObject['$' + argument] :
-        () => gObject[argument]()
+        {}
     };
-    context.$$.dynamic = (value, duration = 200, delay = 0) => {
+    context.$$.currentValue = gObject[argument];
+    context.$$.dynamic      = (value, duration = 200, delay = 0) => {
       gObject.animateTo(
         (isArray(value) ? value : [value]).map(v =>
           isObject(v) && 'offset' in v ?
@@ -156,7 +183,7 @@ defineDirective({
         {duration, delay}
       );
     };
-    let value          = evalExpression(expression, context);
+    let value               = evalExpression(expression, context);
     if (argument === 'class') {
       if (isArray(value)) {
         gObject.classList.add(...value.filter(val => !!val));
@@ -280,14 +307,14 @@ function defineDirective ({name, alias, argument, template, execute}) {
   }`
   const check = new RegExp(source, 'i')
 
-  directives.push({
+  directives[name] = {
     name,
     alias,
     argument,
     template,
     execute,
     check
-  });
+  };
 
 }
 
@@ -297,8 +324,9 @@ function defineDirective ({name, alias, argument, template, execute}) {
  * @returns {Object}
  */
 function findDirective (key) {
-  for (const definition of directives) {
-    const match = definition.check.exec(key);
+  for (const name in directives) {
+    const definition = directives[name];
+    const match      = definition.check.exec(key);
     if (match) {
       let argument = match[2];
       return {...definition, argument}
@@ -480,7 +508,7 @@ function install (setup) {
     extendTemplate : {
       defineDirective,
       obtainDirective (name) {
-        return directives.find(directive => directive.name === name);
+        return directives[name];
       }
     }
   });