+++ test/component/template/assets/colors.js
@@ -1,9 +1,0 @@
-function color (value) {
-  if (value >= 75) {
-    return 'blue';
-  }
-  if (value >= 50) {
-    return 'green';
-  }
-  return 'red';
-}